backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Darius1223 <dar.290199@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.112.2"}
SQLAlchemy = {extras = ["asyncio"], version = "^2.0.32"}
alembic = "^1.13.2"
uvicorn = "^0.30.6"
pydantic = "^2.8.2"
pydantic-settings = "^2.4.0"
psycopg2-binary = "^2.9.9"
structlog = "^24.4.0"
asyncpg = "^0.29.0"
pyjwt = "^2.9.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
sqlmodel = "^0.0.21"
python-multipart = "^0.0.9"
starlette-admin = "^0.14.1"
dependency-injector = "^4.42.0"
redis = "^5.0.8"
aiosmtplib = "^3.0.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.7"
mypy = "^1.11.1"
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
